refactor: prepare lib for release
