+++ Cargo.toml
@@ -2,6 +2,20 @@
 name = "iced-lenia"
 version = "0.1.0"
 edition = "2024"
+description = "Particle Lenia simulation in Rust using iced"
+license = "MIT"
+repository = "https://github.com/lmmx/iced-lenia"
+readme = "README.md"
+keywords = ["iced", "gui", "lenia", "simulation"]
+categories = ["gui", "development-tools"]
+
+[lib]
+name = "iced_lenia"
+path = "src/lib.rs"
+
+[[bin]]
+name = "iced-lenia"
+path = "src/main.rs"
 
 [dependencies]
 iced = { version = "0.12", features = ["async-std", "canvas"] }
@@ -12,3 +26,17 @@ ndarray-rand = "0.15.0"
 rand = "0.9.0"
 rand_distr = "0.5.1"
 rayon = "1.10.0"
+
+[profile.dev]
+opt-level = 0
+lto = false
+codegen-units = 256
+panic = "abort"
+debug = false
+strip = "debuginfo"
+
+[profile.release]
+opt-level = 3
+lto = true
+codegen-units = 1
+strip = true
